3-print_all.c

+++ 0x10-variadic_functions/3-print_all.c
@@ -29,7 +29,7 @@ void print_all(const char * const format, ...)
 				check_stat = 0;
 				break;
 			case 'c':
-				printf("%c", va_arg(spc, float));
+				printf("%c", va_arg(spc, int));
 				check_stat = 0;
 				break;
 			case 's':
